Switch K batching knob from chunks to bytes; uncap MAX_BATCH_EPOCHS
The prior auto-K formula derived epochs-per-batch from `target_batch_chunks`
and clamped to `MAX_BATCH_EPOCHS=128`. For benches with small chunks and one
chunk per epoch (e.g. `smallepoch_single`) the clamp pinned batch payloads
at ~2 MiB, leaving D2H latency-bound (~1.2 GiB/s) instead of
bandwidth-bound. Expressing intent in bytes, and lifting the ceiling, lets
compress/D2H see enough work to amortize.

Changes:
- Config: `target_batch_chunks` (u32) -> `target_batch_bytes` (u32); widen
  `epochs_per_batch` to u32.
- Solver: `compute_epochs_per_batch` derives K from bytes_per_epoch summed
  across LOD levels. Drops pow2 rounding. Removes `MAX_BATCH_EPOCHS`. Keeps
  a soft `AUTO_K_MAX = 65536` as a backstop for pathological tiny-chunk
  configs (user-set `epochs_per_batch` is not capped).
- Heap-alloc the K-sized scratch arrays that previously were
  `[MAX_BATCH_EPOCHS]` inline: per-flush-slot masks, per-stage scratch, CPU
  pipeline scratch. Multiarray descriptors own their own masks; bind/unbind
  just swaps pointers.
- Collapse per-epoch `pool_events[K]` into a single batch-level
  `pool_ready`: all scatters run on the compute stream in order, so an
  event recorded after the last scatter subsumes the per-epoch waits.
- Bench default `target_batch_bytes = 512 MiB`.
- Docs updated to match.

Smallepoch bench throughput: 1.22 GiB/s -> 4.21 GiB/s.

+++ tests/test_gpu_helpers.h
@@ -23,7 +23,7 @@ int
 make_test_config(struct tile_stream_configuration* config,
                  struct dimension* dims,
                  struct codec_config codec,
-                 uint8_t epochs_per_batch);
+                 uint32_t epochs_per_batch);
 
 // Fill one epoch of chunk pool on device.
 // Each chunk t gets all elements set to fill_fn(t).
